+++ src/test/scala/edu/ie3/simona/agent/participant/ParticipantAgentSpec.scala
@@ -48,7 +48,10 @@ import edu.ie3.simona.ontology.messages.{
 }
 import edu.ie3.simona.service.Data.PrimaryData.{ActivePower, ActivePowerExtra}
 import edu.ie3.simona.service.DataTimeType
-import edu.ie3.simona.service.results.ResultServiceProxy.ExpectResult
+import edu.ie3.simona.service.results.ResultServiceProxy.{
+  ExpectResult,
+  NoResult,
+}
 import edu.ie3.simona.test.common.UnitSpec
 import edu.ie3.simona.util.TickUtil.TickLong
 import edu.ie3.util.TimeUtil
@@ -97,7 +100,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val scheduler = createTestProbe[SchedulerMessage]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
         // no additional activation ticks
@@ -254,7 +258,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val scheduler = createTestProbe[SchedulerMessage]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
         // with additional activation ticks
@@ -413,7 +418,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val scheduler = createTestProbe[SchedulerMessage]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
         val service = createTestProbe[ServiceMessage]()
 
@@ -613,7 +619,7 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
         participantAgent ! Activation(18 * 3600)
 
         // nothing should happen, still waiting for secondary data...
-        resultProxy.expectNoMessage()
+        resultProxy.expectMessage(NoResult(MockParticipantModel.uuid, 54000))
         scheduler.expectNoMessage()
 
         participantAgent ! DataProvision(
@@ -698,7 +704,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val scheduler = createTestProbe[SchedulerMessage]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val service = createTestProbe[ServiceMessage]()
 
         // no additional activation ticks
@@ -938,7 +945,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val em = createTestProbe[FlexResponse]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
         // no additional activation ticks
@@ -1131,7 +1139,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val em = createTestProbe[FlexResponse]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
 
         // with additional activation ticks
@@ -1396,7 +1405,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val em = createTestProbe[FlexResponse]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val responseReceiver = createTestProbe[MockResponseMessage]()
         val service = createTestProbe[ServiceMessage]()
 
@@ -1845,7 +1855,8 @@ class ParticipantAgentSpec extends ScalaTestWithActorTestKit with UnitSpec {
 
         val em = createTestProbe[FlexResponse]()
         val gridAgent = createTestProbe[GridAgent.Message]()
-        val resultProxy = createTestProbe[ResultEvent | ExpectResult]()
+        val resultProxy =
+          createTestProbe[ResultEvent | ExpectResult | NoResult]()
         val service = createTestProbe[ServiceMessage]()
 
         // no additional activation ticks